tests/wpt/mozilla/tests/mozilla/characterdata.html

<html>
  <head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <p>This is the character data</p>
  </body>
  <script>
  test(function() {
    var a = document.getElementsByTagName('p')[0].childNodes[0];
    a.data = "This is the character data test, append 資料，更多資料";

    // replace test non-utf8
    a.replaceData(33, 6, "other");
    assert_equals(a.data, "This is the character data test, other 資料，更多資料");
    // replace test non-utf8
    a.replaceData(44, 2, "文字");
    assert_equals(a.data, "This is the character data test, other 資料，更多文字");

    // substring test non-utf8
    assert_equals(a.substringData(12, 4), "char");
    // substring test utf8
    assert_equals(a.substringData(39, 2), "資料");
  });
  </script>
</html>
